requirements.txt
-i https://pypi.org/simple
alembic==1.12.0; python_version >= '3.7'
aniso8601==9.0.1
azure.cognitiveservices.speech==1.31.0
blinker==1.6.2; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
cloudinary==1.34.0
datetime==5.2
flask==2.3.3
flask-cors==4.0.0
flask-migrate==4.0.4
flask-restful==0.3.10
flask-sqlalchemy==3.0.5; python_version >= '3.7'
greenlet==2.0.2; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==21.2.0
honcho==1.1.0
importlib-metadata==6.8.0; python_version < '3.10'
importlib-resources==6.0.1; python_version < '3.9'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
mako==1.2.4; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
packaging==23.1; python_version >= '3.7'
psycopg2==2.9.7
python-dotenv==1.0.0
pytz==2023.3
setuptools==68.1.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.20
typing-extensions==4.7.1; python_version >= '3.7'
urllib3==1.26.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.3.7; python_version >= '3.8'
zipp==3.16.2; python_version >= '3.8'
zope.interface==6.0; python_version >= '3.7'
